+++ src/index.js
@@ -5,7 +5,6 @@ module.exports = (...userConfig) => (env, argv) => {
   return webpackMerge.merge(
     require('./base.js'),
     !Helpers.isHmr() ? require('./recipes/manifest.js') : {},
-    !Helpers.isHmr() ? require('./recipes/cleanOutputDir.js') : {},
     require('./recipes/devServer.js'),
     require('./recipes/extractVendor.js'),
     require('./recipes/dotEnvPlugin.js'),