+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 
 ## Unreleased
 
-* Output path has been changed from `./public` to `./public/dist`
+* :warning: Output path has been changed from `./public` to `./public/dist`
+* Upgrade `webpack-dev-server`
+* Tweak `babel-loader` and `terser` options
 
 ## [0.30.0](https://github.com/ankurk91/laravel-bundler/compare/0.29.3..0.30.0)